fix: resolve pre-existing clippy and CI gate failures across test files
Fix clippy warnings/errors and CI gate failures that were pre-existing
on master but surfaced when running ci-gate with --all-targets:

- Add #[allow(clippy::field_reassign_with_default)] to test files using
  Default::default() followed by field assignments
- Add #[allow(clippy::panic)] for test files with legitimate panic! usage
- Add #[allow(clippy::expect_used)] for test files using expect()
- Add #[allow(clippy::approx_constant)] for test values near PI
- Add #[allow(clippy::assertions_on_constants)] for constant assertions
- Add #[allow(clippy::overly_complex_bool_expr)] for tautological exprs
- Fix unused imports, unused variables, and unused functions via clippy --fix
- Fix collapsible if-let in perl-path-security test
- Fix map_identity in perl-lsp-code-actions test
- Fix redundant let rebinding in perl-lsp-providers test
- Fix unresolved import of format_command_args (moved to perl-dap-shell)
- Replace unwrap() in doc comment with pattern match to fix ratchet check
- Remove clone_on_copy warnings in perl-workspace-discovery
- Remove stale agent worktrees blocking ci-check-no-nested-lock

+++ crates/perl-subprocess-runtime/tests/extended_unit_tests.rs
@@ -393,7 +393,7 @@ fn mock_runtime_alternating_success_failure() -> Result<(), Box<dyn std::error::
         if i % 2 == 0 {
             runtime.add_response(MockResponse::success(format!("ok{}", i).into_bytes()));
         } else {
-            runtime.add_response(MockResponse::failure(format!("err{}", i).into_bytes(), i as i32));
+            runtime.add_response(MockResponse::failure(format!("err{}", i).into_bytes(), i));
         }
     }
